+++ script/places,js
@@ -1,0 +1,3 @@
+export const initialCards = [
+    
+]
