+++ script/utils.js
@@ -1,0 +1,4 @@
+export function closePopup(popupElement) {
+    popupElement.classList.remove("popup_active");
+  }
+
